sprint 01 solution

+++ sprint1/problems/final_task/solution/src/main.cpp
@@ -16,8 +16,8 @@ namespace {
 
 // Запускает функцию fn на n потоках, включая текущий
 template <typename Fn>
-void RunWorkers(unsigned n, const Fn& fn) {
-    n = std::max(1u, n);
+void RunWorkers(size_t n, const Fn& fn) {
+    n = std::max(static_cast<size_t>(1), n);
     std::vector<std::jthread> workers;
     workers.reserve(n - 1);
     // Запускаем n-1 рабочих потоков, выполняющих функцию fn
@@ -37,7 +37,7 @@ int main(int argc, const char* argv[]) {
     try {
         // 1. Загружаем карту из файла и построить модель игры
         model::Game game = json_loader::LoadGame(argv[1]);
-        //model::Game game = json_loader::LoadGame("../../data/config.json");
+        //model::Game game = json_loader::LoadGame("../../data/config.json"); // for debug
 
         // 2. Инициализируем io_context
         const unsigned num_threads = std::thread::hardware_concurrency();
